changed all event handlers to jquery bindings
Previously, event handlers were defined as properties of the relevant elements in the XSLT templates.
This required a few generic functions in Numbas.controls, such as doPart, which would associate part path strings with part objects.
It's also bad practice to define handlers like that these days, because the browser has to eval them, and you're restricted to one-liners.

Anyway, event handlers are now all bound in display.js.