beginning to think about follow-on marks
The student's answers to question parts get saved in variables of the form $pPgG.
So, for example, their answer to the second gap in the first part is saved as $p0g1.

Currently only JME parts can access these variables.

Really, what's needed is a facility for recompiling all the question's variables when one of them changes.