changed the steps marking algorithm.
A problem arose with steps marking today, and there was no
solution with the current steps marking algorithm.
Stacey set up a problem as follows:
- The part is worth 2 marks if you get it right.
- Steps are available, but looking at them will cost you a mark.
- There is a single marked step, worth 0.5 marks.

So what should happen is as follows:
- Getting the part right without looking at steps gets you 2 marks
- Getting the part right after looking at steps gets you at most 1 mark
- Getting the final solution wrong but the step right gets you half a mark.
- You should get more points for getting the final solution but not the
  steps, than for getting the steps but not the final solution.

With the current algorithm, this is four constraints on three variables,
and there's no solution.

Additionally, I think that if, after revealing steps, you get the final
solution right, it shouldn't matter what you wrote in the steps box.
This way, students who just needed a little hint don't have to plod
through putting intermediate results in.

So this is the formula I came up with:

Score awarded = (partMarks - stepsPenalty) * partCredit +
				(1-partCredit)*stepsMarks*stepsCredit

partCredit and stepsCredit are numbers in the range [0,1] representing
how correct the student's answers to the final solution and the steps,
respectively, were.

With this formula, setting the following values for the part gives
the desired behaviour:
partMarks = 2
stepsPenalty = 1
stepsMarks = 0.5

The fourth constraint is satisfied because partMarks is greater than
stepsMarks.