rewrote simplification settings parsing
the rules used when simplifying an expression are now defined by inclusion/exclusion of named sets

A new property of the exam settings object allows you to define both new rules and new rulesets.
It works like so:

```
settings: {
	feedback: {...}
	navigation: {...}

	rulesets: {
		surdFractions: [
			{
				pattern: 'a/sqrt(b)'
				conditions: ['a=b']
				result: 'sqrt(a)'
			}
		]

		cancelUnit: [unitFactor,unitPower,unitDenominator]
		cancelZero: [zeroFactor,zeroTerm,zeroPower,zeroBase]

		cancelFactors: [cancelUnit,cancelZero,!zeroTerm]
	}
}
```

Each object in rulesets is an array of new rule definitions or of names of existing sets.
Sets are constructed by reading left-to-right through the array, adding in named sets and new rules, or removing every rule belonging to a named set with an exclamation mark in front of it.

So in the given example, cancelFactors consists of every rule in cancelUnit and cancelZero, except the zeroTerm rule.

The syntax is the same for specifying rules to be used in the LaTeX \simplify command, eg.

```
\simplify[cancelFactors,surdFractions,!zeroBase]{ ...expr... }
```

Defining commonly-used rulesets at the top of the exam should make using \simplify a lot less complicated.

This will need documenting when it gets to the master branch, @anthonyyoud.