really really simple extension to send all SCORM data to a remote server
for Mike Robinson at Sheffield Hallam

To use this, you MUST change the variable 'address' in extensions/simplesave/simplesave.js to the address of a script which will receive the data.

All scorm data model elements will be sent to that address, as JSON objects in the following format:
{
	name: <student's name>,
	exam: <exam name>,
	element: <data model element>,
	value: <value>
}

It's up to you to write the script which receives this data. Something simple in PHP which writes to a sqlite database shouldn't take long.

The most important elements are:
  cmi.score.raw - total marks for exam
  cmi.objectives.N.id - name of part
  cmi.objectives.N.score.raw - marks for individual parts
  cmi.interactions.N.learner_response - student's answer for part

Suspend/resume happens in local storage, and no data are requested from the server, so students must start and finish the assessment on the same computer.
There are ZERO security measures, so do not use this for high-stakes assessments.