fix display of powers of complex numbers
because exponentiation of complex numbers is calculated by working out arg(z) and abs(z), a little bit of floating point inaccuracy is introduced.
the niceNumber display function was just checking n.im==0, which failed in this sort of situation, so I've made it round to the same accuracy as the rest of the niceNumber routine.

fixes #72