move PretendLMS to an extension
localStorage suspend/resume now only happens if you load the pretendlms extension.
This is because IE8 claimed not to have localStorage, and also because debugging exams is hard when it keeps trying to resume a session started with a different version of the exam.
This does not affect SCORM communication with a proper LMS, eg Blackboard.