annotate variable names to provide information about what category they're in.
There was no way of telling the system that a variable represents, for example, a vector or a matrix. These have their own notation, so it needed to be possible.

Putting 'v:' efore a variable name will mark it as a vector. Similarly, 'm:' marks a variable as a matrix.

So:
	'v:x' will be TeXed as '\mathbf{x}'
	'm:x' will be TeXed as '\mathrm{X}'

The annotations shouldn't conflict with anything on the evaluation side, so use them at will. You can use them throughout, or only in \simplify{} expressions, it makes no difference.

Suggestions for other display styles welcome.