don't bind variable definitions before evaluation
viz #96

In jme-variables.js, syntax trees for each variable definition are generated at the first step.
Then the dependency tree is calculated, and variables are evaluated in order of dependency.
So the eventual type of a variable is unknown until it has been evaluated.
Which means that code shouldn't be bound to the syntax tree until all of its dependencies have been evaluated!
So now variable definitions are compiled with the 'notypecheck' flag set to true, which means code is bound only when they are evaluated.

This, of course, calls into question the point of having the pre-binding thing at all. Needs thinking about, and maybe yet another compilation method!