modified textile.js to deal with plaiun blocks better
viz #118

Most of the time, when jme.subvars calls textile, it is converting inline text which shouldn't be enclosed in <p> tags.
By default, textile wraps plain blocks (blocks which aren't HTML or declared to have any type, like `h1. header`, in <p> tags.
I got round this by prepending a space character to strings before they are passed through textile.
But when you actually *do* want a block, like a table, this causes problems.
So I've added a nowrapPlainBlocks option to the textile parser, which tells it not to enclose plain blocks in <p> tags, but behaves as normal on everything else.