define JME functions with javascript in the .exam file
Instead of giving a 'definition' property, if you give one called 'javascript', you can define what a function does with javascript.

	functions: {
		sum: {
			parameters: [[a,list]]
			type: number
			javascript: """
				var sum = 0;
				for(var i=0;i<a.length;i++)
				{
					sum += a[i].value;
				}
				return sum;
			"""
		}
	}

All of the function's parameters get passed through with the names you give.
You can either return simple javascript data type (in the example above a javascript number is returned, not a JME one) and it will have the appropriate JME constructor applied to it, or you can return an instance of jme.types.TWhatever yourself, if that's more convenient.

Note that JME lists are a bit awkward because they contain JME data types, not javascript ones. In the above example, I used the `.value` property to get the numerical value of a list element.

Vectors are easier to work with - they are simple javascript arrays of numbers.