'op' and 'verb' annotations; 'verbatim' function; properly texify listval
array indexing is LaTeXed halfway sensibly.

'op' is a shorthand to wrap a name in the \operatorname command, so it is displayed using a non-italic font.
eg
	op:E(x)

'verb' (or 'verbatim') is a sort of dummy annotation so that names like `e`, `i` and `pi` don't get interpreted as their respective constants by the compiler.

I've also added a 'verbatim' (or 'verb' for short) function for display, which you can use to insert a string of LaTeX into an expression which otherwise needs simplifying, for example with plus or minus signs.
eg:
	\simplify{ {a}verb('E[X]') + {b}verb('E[y]') }

That's TeX-within-JME-within-TeX. God help us.