rejig how scope is passed around the JME functions
instead of separate `variables` and `functions` objects, a single `scope` object is passed around.
This encapsulates the concept of an "evaluation environment" in which an expression is evaluated, with no reference to javascript's global scope.
Rulesets are now also part of the scope, because a JME expression containing a string with a \simplify command needs to know what rulesets to use.
Up until now, it had just loaded Numbas.exam.rulesets, which causes portability problems.

This is the first step to moving rulesets into questions and out of the parent exam.

Next task is definitely unit tests, because who knows what this broke.