when evaluating a list, return a new TList object
had a case where a user-defined JME function returned a list. When the function was first evaluated, the list was evaluated and its `value` property was filled in.
That value property stuck to the TList object in the function definition, rather than being a new object.
I've changed the jme.evaluate code so a new object is created.

fixes #152