fix siground on negative numbers with a 5 digit on the end
When the most-significant non-significant digit in a negative number (the one you use to decide how to round) is 5, math.siground used to make the wrong decision.
I removed the sign from the number in order to work out the power of 10 so I could get the number in an appropriate form to use javascript's built-in function.
Unfortunately, javascript's built-in function Math.round always has positive bias.
I've fixed it so that Math.round is always given a number of the correct sign.