don't create a new scope for every function application; speed up Array.merge
jme.evaluate was creating a new Scope object for every single function application. As this involves copying, sorting and merging three very big objects, it was very slow.
Particularly on IE8, which seems to be very slow at accessing object properties.
Because of this change, Numbas.jme.evaluate now MUST be passed a Scope object, or it won't see any of the builtin functions.

While I was finding that, I sped up the Array.merge function by sorting and removing adjacent duplicates, not the O(N*N) method I was using before.