tidy up spacing on examparser.printdata on multiline strings
It was always adding newlines and tabs to multiline strings.
Now made it so there's exactly one newline at the start of a string, and exactly the right number of tabs at the start of each line.

Assuming that the exact number of newlines at the start and end of a string will never matter.