don't apply simplification rules to student's JME part answer
for reasons known only to my past self, all the simplification rules were being applied to the student's answer when marking a JME part.
If the answer was an uncancelled fraction, eg 6/12, the simplifyFractions rule would cancel it to 1/2, meaning the max length restriction didn't work.
Now, no rules are applied.
fixes #163