x isa "complex" returns true only if x is n*i
x isa "complex" used to evaluate x and return true if the result is complex. This was bad because something like `3/(2+i)` returns true, but for the purposes of simplification rules shouldn't count as a complex number.
Now it returns true only if x is a single complex number token.
This might have repercussions elsewhere. The simplification system is enormously fiddly.