JME: add HTML data type; subvars works on the DOM
There is a new HTML data type in the JME system, for holding DOM nodes.

The question.subvars method now operates on DOM nodes. The intention is that functions will be able to create dynamic graphics which will go straight into the page, so they preserve things like animations and event bindings.
That can't quite happen yet because everything still goes through XSLT transformations, so it needs to wait for the knockout branch.