add seedrandom.js
seedrandom.js provides a replacement for Math.random which can be seeded. This means I can use a single number as an ID for a particular exam/session, so it can be recreated at any time.
I'm just using it in the worksheet theme at the moment.
It might be a good idea to save the seed for SCORM exams, so we can get round Blackboard not saving the correct answers by just regenerating the exam ourselves.