double-escape characters in the whitespace regexp
I thought I could get away with single-escaping the characters in the definition of the whitespace regexp, because javascript's string literal would convert them to the appropriate characters before they went into the RegExp constructor.
Well, guess which browser didn't parse "\v" as character 11 until its 9th version?

The characters are now double-escaped, so the escapes are handled by the RegExp constructor, which IE8 seems to have a more correct implementation of.

fixes #222