fix question HTML
I had forgotten that if I'm going to keep one copy of a question's HTML for the entire session, I need to not delete it when you change question.

A side-effect of this is that unsubmitted answers persist between question changes, which is something students have been asking for. I need to show unsubmitted answers graphically though, because an answer changed after submitting might look like it's been marked incorrectly.

fixes #238