(almost) all jQuery gone!
The JME preview is *so* much simpler in knockout.
All the jQuery event-binding and DOM manipulation that can be stripped out from display.js has been, apart from the question carousel - not sure if it's worth the effort to make a KO version.

Also moved the Reveal button down to the bottom, next to "Try another". I've been meaning to do that for a while.