fix varnamesAgree for marking JME parts
When a long variable name is used:
`if( (name=array1[i][0])!='$' ... )` set `name` to the first character of the variable name.
Then when the next bit checks if array2 contains `name`, it only looks for the first character.

This changes the line to the more correct
`if( (name=array1[i])[0]!='$' ... )`

fixes #266