only show "submit part" button when answer is changed.
Issue #267 asks for some way of showing the student that the feedback they received for a previously submitted answer might not apply after they've changed their answer.
I'm not keen on hiding the feedback after the answer is changed, because it's still the recorded feedback until the student resubmits.
I thought about making the "submit part" button a bit more attention-grabbing when the answer changes, to prompt the student to submit, but in fact the button doesn't even need to be there the rest of the time.
This commit makes the button invisible until the student enters an answer, or changes a submitted answer.