update JSXGraph extension
The old method of inserting graphs in questions doesn't work with the new theme.
Instead, there's now a function Numbas.extensions.jsxgraph.makeBoard(width,height,options) which creates a div and inits a board in it. The JSXGraph Board object is in div.board.
This also updates jsxgraph to v0.97