trigger 'question-html-attached' event
Javascript functions don't get told which question they belong to, because of the way the scope system works. Additionally, the question HTML doesn't exist at the time the function is called.
To get round this, QuestionDisplay throws an event 'question-html-attached' to the document root once it has finished creating the HTML and subbing in variables.
Functions can listen add hooks to listen to this event with `$(document).on('question-html-attached,function(event,question,questionDisplay){...})`, and do any bindings there.