IE fixes
VML doesn't like being moved in IE, so we now use JSXGraph's canvas renderer.
Change X-UA-Compatible string in the question theme to IE9.
'question-html-attached' event is now triggered on body, because IE can't trigger events on document.
JSXGraph triggers a board.fullUpdate when html is attached, so IE re-renders it.