warn when leaving unsubmitted changes; review screen shows SUBMITTED answers.
When you try to move away from a question with unsubmitted changes, an alert box appears telling you to check your answers and submit. It won't let you leave until everything is submitted.
If the question hasn't previously been answered, the normal warning message is shown, and if set to "warn if unattempted", the student can leave the question with unsubmitted answers. I left this in because some students like to leave hard questions until last. But they shouldn't be allowed to have unsubmitted changes *after* the first submit.
There's now an "unsubmitted changes (!)" line in the question selector under questions with unsubmitted changes.

On trying to end the exam, the same check is done. The only way you should be able to get the "unsubmitted changes" message here is by changing an answer in the current question before ending the exam. Again, it won't let you finish until you've submitted.

When the exam ends, the default theme now resets all inputs to the last submitted answers. It shouldn't be possible for the review mode to disagree with the normal mode now.

This is for #297