\jmescope affects global state instead of environment
The \jmescope[n] command, which is a hack to give \var and \simplify a JME scope to work with, is now a standalone macro which affects global state instead of taking a TeX argument.
It looks like it works, but I'm worried about race conditions, so I'm going to ask the MathJax people about it.