change "must be an integer" warning; reset score feedback when answer changes
viz #324: The "answer must be an integer" restriction doesn't accept answers like "3.0", which are now considered to belong to \mathbb{R} but not \mathbb{Z}.
This also adds a warning "Your answer is invalid. You must enter a whole number, not a decimal." when the student enters a non-integer.

Separately, but annoyingly in the same commit, the score feedback (score amount and tick/cross icon) resets to the unsanswered state when the student changes their answer, until they submit again. This way, it doesn't look like the student has been awarded marks (or scored zero marks) for a different answer to the one that was marked.