don't do the number representation fix on content nodes
the compiler ran code to fix Python's number representation on all strings. This only needs to run on things which the runtime needs to interpret as numbers, and shouldn't change the display for students.
So makeContentNode doesn't do anything to change the string any more, apart from switching it to unicode if the compiler's running in python2

fixes numbas/editor#207