prefix ops don't need to pay attention to operator precedence
The root of numbas/editor#210 wasn't that prefix ops had the wrong precedence - they should just get pushed straight onto the stack without looking at previous symbols.
A prefix op always applies to the thing on its immediate right, so the shunting-yard algorithm should just put it on the stack.
This changes the check when reading an op symbol so that prefix ops can't move any ops from the stack to the output.