add Numbas.addExtension shortcut; move 'base' to start-exam.js
The call to set up the 'base' dependency is in start-exam.js, so the runtime and the editor can use the same version of numbas.js

Numbas.addExtension is a wrapper around Numbas.queueScript. It takes a short name and list of dependencies, then passes a callback the corresponding Numbas.extensions.x object. It means extensions don't need to worry about the `extensions/x/x.js` path, or setting up a JME scope.