Default theme now has a responsive layout
On small screens (tablets and phones, mainly), the default theme didn't used to leave much space for the question content.

This change makes the theme responsive, so screens less than 980px wide switch to a more vertical layout. The question menu and pause/end buttons move beneath the content area, and the banner is more vertical too. This means the content area can have the full width of the page.

Tested on Chrome, Firefox, IE9, various Android and iOS devices.