do parseFloat at the end of precround
instead of (intPart+round(fracPart*b)/b)
The division followed by an addition introduces a floating point error.
For example, `488/1000` is displayed as 0.488 but is actually stored as slightly smaller number.
If you add 4 to that, JavaScript can't recognise the error any more and displays it as `4.4879999999999995`.
However, the literal `4.488` seems to be displayed OK.
So, calculate the integer part as an integer, and the fractional part as an integer (i.e., multiply it by 10^b and round off), then evaluate
`parseFloat(intPart+'.'+fractPart)`
This way, we return something that JavaScript will definitely display as we expect (unless we're trying to get more than 15 digits of precision)