add precision restrictions for matrix parts
Work pretty much the same way as for numberentry parts - for each cell in the matrix, the correct answer is rounded and the student's answer is rounded. If equal, then OK.
Additionally, if the string the user enters for the cell specifies the wrong precision, a penalty is applied (once for the whole part, not once per bad cell).

The precision-checking code from NumberEntryPart.mark has been split off to Numbas.maths.toGivenPrecision so it can be shared between the two parts.