improve DOMsubvars on lists
When you sub a list into HTML with, for example, `{mylist}`, each list element is rendered as if it was subbed in on its own.
In particular, this means strings get treated nicely instead of being rendered as JME code.

You shouldn't sub in a raw list anyway, because it gets wrapped in square brackets, but this might help with debugging.