add "allow fractions" option to numberentry
If true, the student can enter a fraction (a string of the form "digits/digits")
Any amount of whitespace is allowed.

A few new API methods and properties:

util.isNumber(n,allowFractions) - isNaN combined with a check against the regex for a fraction
util.re_fraction - regex recognising fractions
util.parseNumber(n,allowFractions) - parseFloat, or if that gives NaN, try to parse it as a fraction

Additionally, "infinity" is allowed as an answer for numberentry parts (I thought it already was, but apparently not!)

The display code and marking code now use the same method to check the student's answer is valid.