Fix contentsplitbrackets regex
The "start of maths" regex used by contentsplitbrackets has to capture the character before the maths delimiter so it can check it's not a \, since JS regexes don't have negative lookbehind.
We need to return that character to the plain text bit, instead of keeping it as part of the maths delimiter.

Fixes #370