wrapValue can take a type hint for arrays
matrices, vectors, ranges and sets are all implemented as JS arrays.
jme.wrapValue now takes an optional "typeHint" argument which should be the name of the type you're expecting.
It would be better if this logic was abstracted out more and made configurable by extensions, but it'll do for now.