DOMcontentsubvars: data-jme-visible attribute can specify a condition for showing an element
Set the data-jme-visible attribute on an element to specify when it should be shown. For example, you could choose blocks of text to display to the student based on the value of a random variable.
The value of data-jme-variable should be a JME expression which evaluates to either true or false

see numbas/editor#230