scope.evaluate now takes an optional variables argument
It has always been a faff to quickly evaluate a JME expression on a set of variables in JavaScript - you have to convert the variables to JME types and create a scope containing them before evaluating.
Now the Scope object's evaluate method takes an optional dictionary of variable values, which are automatically wrapped up in JME types.