load student's name and ID from SCORM
Had been meaning to do this for a while - get cmi.learner_name and cmi.learner_id and display them on the results page.

They're stored in exam.student_name and exam.student_id, so another storage method (if one is ever written) could fill those in however it likes.