fix `question` in javascript custom function scope
`question` only worked because it was accidentally put in global scope in the Question constructor! That meant that `question` always referred to the last Question object created.
Now makeFunctions (and makeFunction, and makeJavaScriptFunction) take a parameter `withEnv`, which is a dictionary of local variables for the custom function's scope. At the moment, it just has `question`, pointing to the parent question, in it.