noLeadingMinus rule rewrites -0 to 0
fixes #91

I was going to add this rule to basic, but it would lead to possibly unintended consequences in more complicated expressions. For example, `x+(-0)` would be simplified to `x+0`, while `x-0` wouldn't. I can see why you'd want `x-0` to be different from `x+0`, so this rule is only applied when "noLeadingMinus" is switched on.