don't show part marks and submit button when there's only one part
When a question only has one part, it doesn't make sense to have both a "submit part" and "submit answer" button.
The extra marks display in the part feedback is also redundant, since it'll always show the same as the question feedback.

This commit also tidies up some of the logic around the feedback boxes - the part feedback box is completely invisible when none of the buttons or score feedback are shown, and I've fixed the "pulse" animation so it doesn't get triggered just when you change question.