code to find correct answer for parts is separated out into .getCorrectAnswer
Working towards being able to find a new correct answer by replacing question variables with student's answers to other parts.
part.getCorrectAnswer(scope) now uses scope to evaluate the correct answer to the question, and store it in the settings object.
Anything to do with marking that depends on the scope should be in this method, and calling it witha new scope should update all the settings used by the marking algorithm.

Next step is to add a stage to the marking logic which subs in student answers as defined by the question author, and calculates the error-carried-forward score.