add TExpression type
Value of TExpression is a syntax tree.
When a TExpression is substituted into another expression, it's just a subtree - it's not wrapped in the TExpression object. The substitution is recursive - e.g.

	sub {y:TExpression(x+n), n:2} into `y+1` gives `(x+2)+1`

It exists purely to make the error-carried-forward mechanism easier: the answer to a jme part is a TExpression type.
It could probably be used to do more copmuter algebra things like automatic differentiation and anonymous functions, but that needs a lot of thinking.