add plumbing for error carried forward
Methods to get student's answer as JME (and code to get student's answer in a usable form for javascript abstracted out when appropriate)
This also adds:
 * jme.variables.variableDependants, which finds the variables that need to be recomputed when you change one or more others.
 * Part.errorCarriedForwardScope, which uses the dictionary of variable names -> part identifiers in `this.settings.errorCarriedForwardReplacements` to create a new scope with those variables replaced (and their dependants recalculated). You can then plug that into Part.getCorrectAnswer to recalculate the correct answer for the part.

Recalculating dependent variables is really good - one example I've got is a question which asks you to find sample mean and s.d. of two samples, then compute the two-sample t-statistic. Just by replacing the variables mu1,sd1,mu2 and sd2, the t-statistic is automatically recalculated for that part.
The recomputed variables need to be deterministic - it's no good adding random noise that changes each time you submit an answer! There's no way of checking if a variable is deterministic, so we need to add big heavy warnings to the docs.