apply variable replacements
Variable replacements are applied before the Part.mark method, in Part.submit.
Was concerned that it should be part of the mark method, but all sorts of other stuff happens in Part.submit and we can interpret "Part.mark" as "mark against the current setting for correct answers.

Need to add logic to try the built-in answer first, and then the version with variable replacements, and add a feedback message if the replacement version gives a higher score.