add feedback message about using variable replacements
A message saying "This part was marked using your answers to previous parts" is added to the feedback when variable replacements are used.
"Correct answer" is changed to "Expected answer" in the reveal view, to reduce confusion when the student's answer, marked correct due to adaptive marking, doesn't match the expected answer.
I decided not to show the "expected answer" calculated using variable replacements, so that the expected answers are consistent with each other.