set parts with variable replacements as dirty when referenced parts submitted
Suppose part B replaces a variable with the answer to part A.
When part A is submitted, the marking for part A might change, so we need to ask the student to resubmit part A.
When part B is submitted, it stores a reference back to itself in part A. Then when part A is submitted, it alerts every part which it has a back-reference to that it needs to be resubmitted.
At the moment, this just consists of setting the other parts as dirty so the submit button shows again and the student is prevented from leaving until they resubmit.
There should be some feedback to explain what's going on. Maybe it should be a warning on the referenced parts saying something like "because your answer to part N changed, and the marking for this part depends on your answer to part N, you must resubmit this part."

Automatically resubmitting parts is a bad idea because it can lead to all sorts of bubbling and infinite loop problems. It's also quite opaque to the student.