separate out parts into separate files
Each part type's code now lives in its own file.
There are quite a few now, and I want to properly modularise the parts stuff so eventually extensions will be able to add new part types.

Additionally, this separates out the built-in JME functions from the main JME parse/evaluate code.
I've been meaning to do this for a while, so only stuff directly to do with the parsing/evaluating steps is in jme.js