mark parts with 0 marks, and show sensible feedback
Until now, parts with 0 marks were not marked. This was a terrible way of achieving the behaviour "information only parts don't do any marking".
You might want to have a part with zero marks for purely formative purposes, or as part of a clever adaptive marking scheme.
This commit adds a property, `doesMarking`, and a method, `hasStagedAnswer`, to parts to more clearly define what can be done.
When a part is submitted, this happens:
 * if doesMarking is true, then try to mark:
   * if hasStagedAnswer() returns false, show an error and return.
   * otherwise, do the marking
 * otherwise, do nothing

The tick or cross feedback now follows the credit awarded for a part, not the score (which is credit*marks available).
This means that parts with 0 marks will now show a tick or a cross if you get them correct/incorrect.
Multiple response parts with 0 marks will show a tick if there are no reasons to reject the student's choice - for example, the correct numbers of choices and answers are ticked.