improve the rendering of names
This cleans up the logic for Numbas.jme.display.texName.
"Mathsy" names are one letter, followed by a string of digits, or an underscore and a string of digits or at most two alphanumeric characters, and followed by any number of primes. They're rendered in the normal maths font.
Greek letter names are replaced with the appropriate macro.
"Non-mathsy" names are treated as the kind of thing you'd type in a computer, so wrapped in \texttt if they're variable names, or in \operatorname if they're function names. \operatorname escapes underscores so they're displayed and don't cause errors.

Fixes #410