don't round numbers off when saving variables values to suspend data!
The SCORM storage code to save the suspend data for a question runs treeToJME on each question variable to get a string representation.
This by default runs numbers through math.niceNumber, which rounds off to 10 decimal places. Clearly this is no good!

This commit adds a `settings` object to the jmeRationalNumber and jmeRealNumber functions, which allows you to ask them not to use niceNumber. The SCORM storage code uses this flag to make sure the raw numbers are stored.