when evaluating a function, don't save the found funcObj on the token object
jme.evaluate used to keep the funcObj it found in tok.fn. This is bad because if you evaluate the token again on a set of arguments that has no corresponding funcObj, the old funcObj is still there.
This keeps the found funcObj in local scope, so it doesn't persist between evaluations.

This also changes the behaviour of jme.evaluate on an unbound variable: it returns a new TName object, not the original one, so the "unboundName" property isn't saved on the object passed in.