firm up TExpression
TExpression was added while I did the adaptive marking stuff, purely as a way to represent the student's answer to a JME part.

A TExpression token should represent a compiled JME expression. Evaluating it should evaluate the expression. Evaluating a reference to a TExpression should return the TExpression.

jme.substituteTree doesn't try to substitute into a TExpression any more. This might not be right, but the alternative feels very complicated.

This also adds a typeToJME method for TExpressions, which returns `expr(...)`. This function `expr` doesn't exist yet, but maybe it should.