when resuming, don't submit parts until question HTML attached
createPart triggers a submit on the part if the session is being resumed and the part has been answered.
createPart is called before the question HTML is created, so if the marking or validation scripts rely on some html existing, they fail.
This change makes the call to submit run after the question html is attached.

I considered adding a "ready" event for questions, because what we really want to know is that we can interact with the question assuming everything's there, but if parts get submitted at this point and a question author also hooks into this event assuming a fresh question, problems could happen. Using "onHTMLAttached" doesn't solve that problem, so I might want to add a more detailed sequence of event hooks once I've thought about it more.