variable list of choices/answers works when def is HTML
If a string contains html, or the value to use is a THTML type, creating a text node won't do! The logic to create new <choice> or <answer> nodes now goes to a lot more effort to faithfully reproduce HTML.
Because the nodes have to go through the HTML document, any events attached to nodes in THTML values will be lost.