use innerHTML to create nodes from strings when loading answer/choice defs
parseFromString expects all special characters in text nodes to be escaped, and I can't make sure that happens.
Now, we create an HTML node in the page document, set innerHTML to the value of the string, then import it into the XML document and copy the imported nodes's children into the span.